requirements.txt
fastapi>=0.109.2
uvicorn>=0.27.1
heroku-applink>=1.0.0
pytest>=8.2.0
pytest-asyncio==0.23.6
requests>=2.31.0
PyYAML>=6.0.1
pytest-mock>=3.12.0
pytest-cov